correct constants, strings
