+++ app/src/main/java/com/example/compmoviedb/presentation/navigation/NavRoute.kt
@@ -1,9 +1,10 @@
 package com.example.compmoviedb.presentation.navigation
 
-import com.example.compmoviedb.presentation.utils.Constants
+import com.example.compmoviedb.presentation.utils.ScreensConstants
 
 sealed class NavRoute(val route: String) {
-    object Splash : NavRoute(route = Constants.Screens.SPLASH_SCREEN)
-    object Main : NavRoute(route = Constants.Screens.MAIN_SCREEN)
-    object DetailsMovie : NavRoute(route = Constants.Screens.DETAILS_MOVIE_SCREEN)
+
+    object Splash : NavRoute(route = ScreensConstants.Screens.SPLASH_SCREEN)
+    object Main : NavRoute(route = ScreensConstants.Screens.MAIN_SCREEN)
+    object DetailsMovie : NavRoute(route = ScreensConstants.Screens.DETAILS_MOVIE_SCREEN)
 }
