--- app/src/main/java/com/example/compmoviedb/presentation/utils/Constants.kt
@@ -1,18 +1,0 @@
-package com.example.compmoviedb.presentation.utils
-
-object Constants {
-    object Screens {
-        const val SPLASH_SCREEN = "splash_screen"
-        const val MAIN_SCREEN = "main_screen"
-        const val DETAILS_MOVIE_SCREEN = "details_movie_screen"
-    }
-
-    object Keys {
-        const val MOVIE_ID = "movie_id"
-        const val MOVIEDB_BASE_IMAGE_URL = "https://image.tmdb.org/t/p/w500"
-        const val VOTE_AVERAGE = "Среднее количество голосов: "
-        const val RELEASED = "Дата релиза: "
-        const val PRODUCTION_COUNTRIES = "Страна производства: "
-        const val YOUTUBE = "YouTube"
-    }
-}
