+++ app/src/main/java/com/example/compmoviedb/presentation/utils/ScreensConstants.kt
@@ -1,0 +1,14 @@
+package com.example.compmoviedb.presentation.utils
+
+object ScreensConstants {
+
+    object Screens {
+        const val SPLASH_SCREEN = "splash_screen"
+        const val MAIN_SCREEN = "main_screen"
+        const val DETAILS_MOVIE_SCREEN = "details_movie_screen"
+    }
+
+    object Args {
+        const val MOVIE_ID_ARG = "movie_id"
+    }
+}
