+++ app/src/main/java/com/example/compmoviedb/presentation/utils/StringsConstants.kt
@@ -1,0 +1,7 @@
+package com.example.compmoviedb.presentation.utils
+
+object StringsConstants {
+
+    const val EMPTY_STRING = ""
+    const val YOUTUBE = "YouTube"
+}
