+++ app/src/main/java/com/example/compmoviedb/presentation/utils/URLConstants.kt
@@ -1,0 +1,5 @@
+package com.example.compmoviedb.presentation.utils
+
+object URLConstants {
+    const val MOVIEDB_BASE_IMAGE_URL = "https://image.tmdb.org/t/p/w500"
+}
