+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,7 @@
 <resources>
     <string name="app_name">Compmoviedb</string>
+
+    <string name="vote_average">Среднее количество голосов:\u0020</string>
+    <string name="date_released">Дата релиза:\u0020</string>
+    <string name="production_countries">Страна производства:\u0020</string>
 </resources>
